circuit_runner: Add ImprovedRunner
Add the class ImprovedRunner which aims to imporove circuit runner
output from the main pecos circuit_runner. To this end the data
structures MeasurementContainer and Measurement, both inheriting from
dicts are added.

The MeasurementContainer is a dict with added methods:
- MeasurementContainer.to_list()
    returns a list of the values
- MeasurementContainer.simplified()
    compatibility alias for to_list

The Measurement class also inherits from a dict. It stores a single
measurement. It has the added functionality of being able to return the
syndrome of a measurement as a list of bits (0 or 1). The length of the
list is determined by the num_qubits attribute, which is assigned at
initialization. This num_qubits should be derived from the simulator
state. The syndrome can be returned with the Measurement.syndrome()
method.

The ImprovedRunner wraps the run function to improve runner output to
give more consistent results. The standard runner returns an empty set
if no measurements were taken or if a measurement result was 0. This is
confusing. This also goes for faults. The ImprovedRunner aims to improve
the output by giving clearer output. If no measurement was taken or no
faults were added, the runner now returns None in each respective case.
If a measurement was taken, the runner now always returns a
MeasurementContainer containing Measurements at specific ticks (as a
value key pair). Measurements with a 0 measurement result simply store a
0 as their value.