Steane: Add Plaquette
Add Plaquette and Stabilizer datatypes. In the Steane code, stablizers
are defined on plaquettes. These plaquettes consist of 4 vertices, with
qubits on each vertex. Plaquettes define a single stabilizer
(generator). Since plaquettes can define both an X and a Z stabilizer,
it makes sense to define a plaquette and create a stabilizer as an
inheriting class from a plaquette.

The Stabilizer class inherits from the plaquette and additionally
defines a pauli_type. The XStabilizer and ZStabilizer classes inherit
from the Stabilizer class and have the pauli_type fixed to "X" and "Z"
respectively.