Steane: Add Syndrome
The syndrome module implements a namedtuple which can be used to define
a measurement syndrome of a set of stabilizers. It contains the
syndrome_type, top left and right plaquette measurement results.

The SteaneSyndromeDecoder class defines a Look-Up-Table (LOT) decoder
for optimal syndrome decoding. The lot_decoder method decodes the
optimal error correction from the syndrome and returns which qubit to
correct with which pauli_type.