error_gen: Add ErrorGenerator
Add an improved highly customizable ErrorGenerator. The generator allows
for fine control of error generation by providing the user all the tools
needed to define any error type at any type of gate object, be it idle
or gate. Errors are defined using two named tuple types called
ErrorProneGateCollection and IdleErrorCollection. The
ErrorProneGateCollection defines a symbol (used in the pecos error
generator to keep track of this object), error prone gates (ep_gates)
(the gates on which errors may occur), a probability parameter (param),
error_gates (the errors which may occur on the ep_gates), and wheter the
errors occur before and/or after a specific gate. The
IdleErrorCollection does the same as the ErrorProneGateCollection up to
not defining the ep_gates. It operates solely on idle spots.

Some standard ErrorProneGateCollecion objects have been defined for
flipping inits and measurement results.

The GeneralErrorGen class automatically configures all errors based on
the gen_error_prone_gates method. This method is somewhat clunky however
and should probably be revised. Currently the error generator is also
not able to be reconfigured which means that configuration has to be
done strictly through class inheritance and overriding the
gen_error_prone_gates or configure_error_generator methods.