Steane: Add Steane circuits
Defines a BaseSteaneCirc (defining a number of qubits, ancilla and flag
qubit registers). Also defines a simulator which corresponds to the
correct state.

Furthermore many important data types are defined here which have to do
with the base steane circuit. These may better be suited in the
data_types directory however.