error_generator: Fix issue initializing two_qubit errors
Two qubit errors were previously instantiated with the ErrorSet class
from the error generator. This object is meant to be used for single
qubit errors only and the ErrorSetmultiQuditGate class should be used in
this case instead. The fix is to check if any error in an error set is
an iterable and assigning the right classes on an if else basis. A
method by the name of "error_set_from_epgc" has been added to create the
rght ErrorSet class (ErrorSet / ErrorSetMultiQudit) from an epgc object.

Furthermore, code repetition has been reduced by introducing the method
"configure_error_group". This method calls the above mentioned method.
and handles the configuration of an EPGC into an ErrorSet for both
"before" and "after" errors.