Steane.circuits.Steane: Add physical zero init
Add a physical zero initialization (setting all data qubits to 0) which
can be used to initialize a physical zero state. This is usefull for
initializing a state with the method of creating a logical zero state by
running one steane round. Using this method, errors can be applied to
this initialization.