qecc_codes.Steane.circuits.Steane: Make SteaneBaseCirc runnable
Add a runner object to the circuit such that it can be run simply by
calling a run method on the circuit class instance.

SteaneBaseCirc now has an attribute named SteaneBaseCirc._runner which
defines a pecos runner object. In this case, the pecos_toolkit specific
version ImprovedRunner is used.

SteaneBaseCirc.run() takes the optional argument "state" and otherwise
passes all *args, **kwargs to the SteaneBaseCirc._runner.run() method.
If no state is supplied, a new state is initialized from the simulator
attribute. The number of qubits of this simulator instance is defined
by the register size.