qecc_codes.Steane.protocols: Add state decoder
Use classical decoding method instead of codeword detection. The
classical decoder first performs a quantum measurement to get a bit
string after which the bits of each stabilizer are copied analyzed. By
adding the stabilizer bits modulo 2 (in 0, 1 readout representation,
analagous to multiplication in 1, -1 readout representation), three
classical stabilizer syndrome bits are produced. From this, a correction
bit flip may be applied on the original readout bit string. By adding
the bits modulo 2 of any logical operator (for instance 0, 1, 4), a
logical state is derived.

This method can be applied directly onto a state, instead of on a
measurement syndrome.