qecc_codes.steane.circuits.Measurement: Add measure basis
Add optional measurement basis switching for the data qubit readout
circuit. Previously this was hardcoded to the Z basis. However, phase
flips are not visible in the Z basis. It is therefore usefull to be able
to also measure in the X basis. The default basis remains the Z basis.