qecc_codes.steane.circuits.Measurement: Fix F1FTEC measurement
Previously, the measurement basis was swapped based on which plaquette
type was being measured. This is wrong for the way the stabilizer
readouts are implemented, because only X errors propagate into the
circuit and these have to be checked in the Z basis on the flag line.
This issue would have caused non-deterministic readouts on the flag line
during simulations.