qecc_codes.steane.data_types.Syndrome: Fix indexation error
For F1FTEC decoder, there was a fatal indexation error. This caused the
wrong modified correction to be chosen. The indexation was off by 1.
Moreover, if the LOT was modified, it was done so at class level as it
was previously defined as a class attribute. The definition of the LOT
has been placed in the __init__ method such that multiple instances may
have different LOTs based on the flagged circuit.