qecc_codes.steane.protocols: Add circuit swapping pattern
Add a circuit swapping pattern which may be replicated to all protocol
methods. The protocol now has fixed ciruit implementations, however, for
manually placed errors, the protocols may not directly be used. This
requires new implementations of protocol methods if they need to be
tested against manually placed errors.

Another pattern which is currently being considered is turning the
protocol classes into actual classes with an init method. Circuits would
be defined at init such that circuits may be changed in a non-static
way. The protocol instance can then simply be run and analyzed as wished.
This would be a large change and would make scripts using these
protocols incompatible as they rely on the implementations being static
methods.