qecc_codes.steane.protocols: Fix F1FTEC correction from flagged circ
The previous implementation of the correct_from_flagged_circuit method
was faulty for the following reasons:
  - it did not correct both X and Z errors as it should. This removes
    the need to provide the 'stabs' argument as all stabs have to be
    measured anyway.
  - it modified the wrong stabilizer. If an X-stabilizer flags it could
    cause X errors to propagate, which would require a modified Z
    syndrome decoder in the following run. A flagged Z-stabilizer could
    cause Z errors to be propagated, which are measured with the
    X-syndrome readout. Therefore, the modification of the syndrome
    decoder has to be swapped, which it perviously wasn't.

The method could be fixed up a little because the decoder types are
somewhat incompatible. (The standard decoder returns a qubit and
correction type as an int and string, but the F1FTEC decoder returns a
tuple of qubits and a correction type as a string, as it may have
multiple qubits to correct.) One fix which would make the decoders
compatible again is to have them return a suggested correction circuit
instead of correction metadata, or at least a method which does this to
not break backwards compatibility.