error_placer_toolkit: Add a toolkit for manual error placing
The error_placer_toolkit contains a few classes which are usefull for
manually placed errors in circuits. It takes an epgc_list as an error
model and a circuit and can generate all posible error circuits and
error combination circuits. The only thing it cannot do currently is
place errors on dynamically branching circuits. This would require a
looped implementation and measurement checking which is beyond the scope
of the current implementation.