circuit_runner: Fix bug
The circuit runner ran into issues if two measurements took place in one
tick. This caused only the last qudit to be recorded due to faulity
implementation of the assignment of measurement locations. This has now
been fixed with an elegant list comprehension which defines which
locations are being measured. A measurement is only recorded if at least
one measurement gate is present at a single tick.