circuit_runner: Change Measurement.syndrome behaviour
Measurement.syndrome returns a list of binary values with measurement
results. Previously, the syndrome returned 0 even if no measurement
occured for the qubit in question, now these are filled with None. On
locations where measurements took place, the bits are still binary
values representing the result of a 0 measurement or 1 measurement.