qecc_codes.steane.decoders: Add Sequential LOT Decoder
Also moves previous decoders to the decoders directory. This change
still has to be permeated through to other parts of the module that
directly use the decoder, such that all decoders are initialized from
the decoders module instead of data types (where it shouldn't have been
in the first place).