qecc_codes.steane.data_types.RNNDataTypes: new pad_to funcionality
the old pad_to method has been deprecated in favor of padding using
numpy in the 'to_vector' method. Additionally, a unittest for the old
pad_to method has been added. It can still be used but a deprecation
warning is now included when using the old pad_to method as it is
comparatively slow (python loop instead of numpy implementation).

For large datasets this does make a difference!

Closes #4