qecc_codes.steane.decoders.BasicLOTDecoder: Add classical state decoding
Add methods to decode a classical state given an input array. The
classical decoding was previously defined as a protocols.SteaneProtocol
method but it turned out to be usefull in other locations as well.
Therefore, it has been moved into the decoder (as it is truly a version
of decoding using the LOT table approach).