qecc_codes.steane.decoders: Add parity output method
Add a method that directly outputs expected output parity given an input
parity and sequence data. It does this by checking the logical weight of
the suggested compiled correction and comparing it to the input parity.
From this, a final parity prediction is made. This can be checked
against the final measured parity of the dataset, and is useful for
direct comparison against the RNN decoder. If the final predicted parity
is correct, it can be assumed that the sequential LOT decoder is able to
correct it to the propper logical state.